Added maven, changes to UI
