+++ ui/index.html
@@ -4,25 +4,16 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover" />
-  <meta name="Description" content="Put your description here.">
+  <meta name="Description" content="Voxetta, an open-source voice recording data collection appn.">
   <base href="/">
 
-  <style>
-    html,
-    body {
-      margin: 0;
-      padding: 0;
-      font-family: sans-serif;
-      background-color: #ededed;
-    }
-  </style>
-  <title>voxetta-frontend</title>
+  <title>Voxetta</title>
 </head>
 
 <body>
-  <voxetta-frontend></voxetta-frontend>
-
-  <script type="module" src="./src/voxetta-frontend.js"></script>
+  <voxetta-main></voxetta-main>
+  <voxetta></voxetta>
+  <script type="module" src="./src/VoxettaMain.js"></script>
 </body>
 
 </html>
