--- ui/src/voxetta-frontend.js
@@ -1,3 +1,0 @@
-import { VoxettaFrontend } from './VoxettaFrontend.js';
-
-customElements.define('voxetta-frontend', VoxettaFrontend);
